batch model meshes 81 -> 3 draw calls

+++ engine/src/core/Texture2D.hpp
@@ -5,8 +5,8 @@
 namespace R3 {
 
 // clang-format off
-enum class TextureType : uint32{
-    Undefined   = std::numeric_limits<uint32>::max(),
+enum class TextureType : uint8{
+    Undefined   = std::numeric_limits<uint8>::max(),
     Diffuse     = 0,
     Specular    = 1,
     Normals     = 2,
@@ -16,12 +16,12 @@ enum class TextureType : uint32{
 // clang-format on
 struct TextureTypeBits {
     TextureTypeBits() = delete;
-    static constexpr uint32 Undefined = static_cast<uint32>(TextureType::Undefined);
-    static constexpr uint32 Diffuse = static_cast<uint32>(TextureType::Diffuse);
-    static constexpr uint32 Specular = static_cast<uint32>(TextureType::Specular);
-    static constexpr uint32 Normals = static_cast<uint32>(TextureType::Normals);
-    static constexpr uint32 Height = static_cast<uint32>(TextureType::Height);
-    static constexpr uint32 Emissive = static_cast<uint32>(TextureType::Emissive);
+    static constexpr uint8 Undefined = static_cast<uint8>(TextureType::Undefined);
+    static constexpr uint8 Diffuse = static_cast<uint8>(TextureType::Diffuse);
+    static constexpr uint8 Specular = static_cast<uint8>(TextureType::Specular);
+    static constexpr uint8 Normals = static_cast<uint8>(TextureType::Normals);
+    static constexpr uint8 Height = static_cast<uint8>(TextureType::Height);
+    static constexpr uint8 Emissive = static_cast<uint8>(TextureType::Emissive);
 };
 
 class Texture2D {
@@ -45,7 +45,7 @@ public:
 
     void bind(uint8 index);
     auto type() const -> TextureType { return m_type; }
-    auto typeBits() const -> uint32 { return static_cast<uint32>(m_type); }
+    auto typeBits() const -> uint8 { return static_cast<uint8>(m_type); }
 
 private:
     uint32 m_id{0};
